requirements.txt
asn1crypto==0.24.0
attrs==19.3.0
aws-xray-sdk==0.95
beautifulsoup4==4.9.1
blis==0.4.1
boto==2.49.0
boto3==1.9.106
botocore==1.12.106
Bottleneck==1.3.2
catalogue==1.0.0
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
clipster==2.0.2
conda==4.6.7
cryptography==2.5
cycler==0.10.0
cymem==2.0.3
dataclasses==0.7
decorator==4.4.2
docker==3.7.0
docker-pycreds==0.4.0
docutils==0.14
ecdsa==0.13
et-xmlfile==1.0.1
fastai==1.0.61
fastprogress==0.2.3
funcsigs==1.0.2
furl==2.1.0
future==0.17.1
greenlet==0.4.15
humanfriendly==8.2
idna==2.8
imageio==2.8.0
imgaug==0.4.0
importlib-metadata==1.6.1
jdcal==1.4
jmespath==0.9.4
jsondiff==1.1.1
jsonpickle==1.1
jsonschema==3.2.0
kiwisolver==1.2.0
matplotlib==3.2.2
mock==2.0.0
msgpack==0.5.6
murmurhash==1.0.2
neovim==0.2.6
neovim-remote==2.1.0
networkx==2.4
numexpr==2.7.1
numpy==1.16.2
nvidia-ml-py3==7.352.0
opencv-python==4.2.0.34
openpyxl==2.5.4
orderedmultidict==1.0.1
packaging==20.4
pandas==1.0.5
pathlib2==2.3.5
pathtools==0.1.2
pbr==5.1.3
Pillow==6.2.2
plac==1.1.3
plotly==4.8.1
praw==6.4.0
prawcore==1.0.1
preshed==3.0.2
psutil==5.4.7
pycosat==0.6.3
pycparser==2.19
pycryptodome==3.7.3
PyJWT==1.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.6.8
python-dateutil==2.8.0
python-jose==2.0.2
pytz==2017.3
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.21.0
requests-file==1.5.1
retrying==1.3.3
ruamel-yaml==0.15.46
s3transfer==0.2.0
scikit-image==0.17.2
scikit-learn==0.20.3
scipy==1.2.1
Shapely==1.7.0
six==1.12.0
soupsieve==2.0.1
spacy==2.3.0
srsly==1.0.2
thinc==7.4.1
tifffile==2020.6.3
torch==1.3.1
torchvision==0.4.2
tqdm==4.46.1
trains==0.15.1
typing==3.7.4.1
update-checker==0.16
urllib3==1.24.1
wasabi==0.7.0
watchdog==0.10.2
websocket-client==0.55.0
Werkzeug==0.14.1
wrapt==1.11.1
xlib==0.21
xmltodict==0.12.0
zipp==3.1.0
